added functional text boxes for npc dialogue
The text boxes are still a bit buggy. I'm gonna keep working on them, but for now they will definitely work if you copy either of the 2 NPCs (make prefabs out of them) who are currently speaking with text boxes in the sample scene I was working in. To change the dialog you need to use the editable field in the Dialog Manager object